pyproject.toml
[project]
name = "rag-interview"
version = "0.1.0"
description = ""
authors = [
    {name = "Huy Dao Ngoc",email = "dnhuy2802@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.14,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-genai (>=1.24.0,<2.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-google-genai (>=2.1.6,<3.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uuid (>=1.30,<2.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "langgraph (>=0.5.1,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
